refactor: split pkg/runtime/docker/docker.go - sidecar.go + netem.go + iptables.go (issue 3, part C)
